+++ process/elasticproc/factory/elasticProcessorFactory.go
