+++ process/elasticproc/factory/elasticProcessorFactory_test.go
